Extend mkyaffs2image to support setting SELinux security contexts in yaffs2 images.
Modern yaffs2 includes support for extended attributes, and thus can support
SELinux security contexts.  Extend mkyaffs2image with support for looking
up the right security context from the file_contexts configuration and
setting it in the generated image.  This is similar to the existing support
for looking up the UID/GID/mode via android_filesystem_config.h and setting it,
but via configuration rather than defined in a header.

Change-Id: Ic86358987f39bddbb83ee928bbbc8cb213efb752