Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I'm not sure I agree that this is the best way to deal with this issue.  I do understand the problem, and this is the most central place to fix it.  But changing all relative paths to be absolute paths simply seems wrong to an old Unix hacker like me.  Instead, the relative paths should be updated.

However, since there are two different ideas of the root of the tree, I don't see a better way than this to resolve the problem.

Patch-set: 1
Label: Code-Review=+1